@@ -1,19 +1,23 @@
-const cookie = require('cookie');
-const jwt = require('jsonwebtoken');
+// @flow
+
+import cookie from 'cookie';
+import jwt from 'jsonwebtoken';
 const JWT_SECRET = process.env.JWT_SECRET;
 
-const Room = require('./game/models/Room.js');
-const Player = require('./game/models/Player.js');
+import Room from './game/models/Room.js';
+import Player from './game/models/Player.js';
 
-module.exports = class RoomServer {
+export default class RoomServer {
+  rooms: Map<string, Room>;
+  userToRoom: Map<string, string>;
+  io: any;
 
   constructor() {
-    this.rooms = new Map(); // { id: room }
-    this.userToRoom = new Map(); // { userId: roomId }
-    this.io = null;
+    this.rooms = new Map();
+    this.userToRoom = new Map();
   }
 
-  listen(io) {
+  listen(io: any) {
     const namespace = io.of('/game');
     this.io = namespace;
     namespace.on('connection', socket => {
@@ -51,28 +55,6 @@ module.exports = class RoomServer {
         });
       });
 
-      // socket.emit('init', {
-        // userId: socket.userId,
-        // deck: socket.deck,
-      // });
-
-      // socket.on('createRoom', () => {
-        // const roomId = this.generateRoomId();
-        // this.saveRoom(new Room({ id: roomId }));
-        // this.join(roomId, socket);
-      // });
-
-      // socket.on('joinRoom', roomId => {
-        // this.join(roomId, socket);
-      // });
-
-      // socket.on('leaveRoom', roomId => {
-        // this.leave(roomId, socket);
-      // });
-
-      // socket.on('disconnect', () => {
-        // console.log('disconnected');
-      // });
     });
 
   }
@@ -88,60 +70,39 @@ module.exports = class RoomServer {
     return id;
   }
 
-  // createRoom(userId) {
-    // const roomId = this.generateRoomId();
-    // const room = new Room({ id: roomId });
-    // this.rooms.set(room.id, room);
-    // this.userToRoom.set(userId, room.id);
-  // }
-
-  getRoom(roomId) {
+  getRoom(roomId: string) {
     if (!this.rooms.has(roomId)) {
       throw new Error(`Room[${roomId}] is not exists.`);
     }
     return this.rooms.get(roomId);
   }
 
-  // enterRoom(userId, room) {
-    // this.saveRoom(room);
-    // this.io.to(room.id).emit('initRoom', {
-      // userId,
-      // room: room.toJSON()
-    // });
-  // }
-
-  syncRoom(room) {
+  syncRoom(room: Room) {
     this.saveRoom(room);
-    // this.io.to(room.id).emit('syncRoom', room.toJSON());
     this.io.to(room.id).emit('syncRoom', room.toData());
   }
 
-  syncGame(room, action=null) {
+  syncGame(room: Room, action?: { from:number, to:number, target:?number }) {
     this.saveRoom(room);
-    this.io.to(room.id).emit('syncGame', {
-      game: room.game.toData(),
-      action,
-    });
-    // this.io.to(room.id).emit('syncUnits', {
-      // /* FIXME action: unitSeq, from, to, target */
-      // units: room.game.units.map(unit => {
-        // return {
-          // seq: unit.seq,
-          // hp: unit.hp,
-          // cellId: unit.cellId,
-          // isActed: unit.isActed,
-        // };
-      // }).toArray(),
-    // });
+    const { game } = room;
+    if (game) {
+      this.io.to(room.id).emit('syncGame', {
+        game: game.toData(),
+        action,
+      });
+    }
   }
 
-  saveRoom(room) {
+  saveRoom(room: Room) {
     this.rooms.set(room.id, room);
   }
 
-  join(roomId, userId, deck, socket) {
+  join(roomId: string, userId: string, deck: Array<number>, socket: any) {
     try {
       let room = this.getRoom(roomId);
+      if (!room) {
+        throw `room ${roomId} is not exists.`;
+      }
       socket.join(roomId);
       console.log(`join ${userId} to room ${roomId}`);
 
@@ -149,53 +110,82 @@ module.exports = class RoomServer {
         id: userId,
         deck: deck
       });
-      room = room.addPlayer(player);
+      room.addPlayer(player);
       socket.emit('enterRoom', { userId });
       this.syncRoom(room);
 
       socket.on('getBattleReady', () => {
         let room = this.getRoom(roomId);
-        this.syncRoom(room.getBattleReady(userId).mightStartGame());
+        if (room) {
+          this.syncRoom(room.getBattleReady(userId).mightStartGame());
+        }
       });
       socket.on('selectUnits', ({ list }) => {
         let room = this.getRoom(roomId);
-        this.syncRoom(room.selectUnits(userId, list).mightEngage());
+        if (room) {
+          this.syncRoom(room.selectUnits(userId, list).mightEngage());
+        }
+      });
+      socket.on('syncGame', () => {
+        let room = this.getRoom(roomId);
+        if (room) {
+          const { game } = room;
+          if (game) {
+            socket.emit('syncGame', {
+              game: game.toData(),
+            });
+          }
+        }
       });
       socket.on('act', ({ from, to, target }) => {
         let room = this.getRoom(roomId);
-        this.syncGame(
-          room.actInGame(userId, from, to, target).mightResetPlayers(),
-          { from, to, target }
-        );
+        if (room) {
+          const changes = room.actInGame(userId, from, to, target);
+          room.mightResetPlayers();
+          this.syncGame(
+            room,
+            { from, to, target, changes }
+          );
+        }
       });
       socket.on('endTurn', () => {
         let room = this.getRoom(roomId);
-        this.syncGame(room.endTurn(userId).mightResetPlayers());
+        if (room) {
+          this.syncGame(room.endTurn(userId).mightResetPlayers());
+        }
       });
-
+      socket.on('returnRoom', () => {
+        let room = this.getRoom(roomId);
+        if (room) {
+          socket.emit('syncRoom', room.toData());
+        }
+      });
     } catch (e) {
+      console.log(e);
       socket.disconnect();
       return;
     }
   }
 
-  leave(userId) {
+  leave(userId: string) {
     if (!userId) {
       return;
     }
     const roomId = this.userToRoom.get(userId);
-    try {
-      let room = this.getRoom(roomId);
-      room = room.leave(userId);
-      if (room.players.count() == 0) {
-        this.rooms.delete(roomId);
-      } else {
-        this.syncRoom(room);
-      }
-    } catch (e) {
+    if (!roomId) {
       return;
+    }
+    let room = this.getRoom(roomId);
+    if (!room) {
+      return;
+    }
+    room.leave(userId);
+    if (room.players.size == 0) {
+      this.rooms.delete(roomId);
+    } else {
+      this.syncRoom(room);
     }
     this.userToRoom.delete(userId);
   }
 
-};
+}
