@@ -1,88 +1,104 @@
-const Game = require('../../models/Game.js');
-const Unit = require('../../models/Unit.js');
-const Immutable = require('immutable');
-
-beforeAll(() => {
-  require('../../data').init();
-});
+import Game from '../../models/Game.js';
+import Unit from '../../models/Unit.js';
 
 describe('Game', () => {
-  describe('toData()', () => {
-    test('return object', () => {
-      const game = new Game({
-        cost: 1,
-        fieldId: 2,
-        turnCount: 3,
-        turn: 4,
-        isEnd: true,
-        winner: false,
-      });
 
-      expect(game.toData()).toEqual({
-        cost: 1,
+  describe('new', () => {
+    test('create Game instance', () => {
+      const actual = new Game({
         fieldId: 2,
-        turnCount: 3,
-        turn: 4,
-        isEnd: true,
-        winner: false,
-        units: [],
-      });
-    });
-  });
-
-  describe('restore()', () => {
-    test('return Game instance', () => {
-      const actual = Game.restore({
+        state: {
+          turnCount: 3,
+          turn: false,
+          isEnd: true,
+          winner: false,
+        },
         units: [
           { unitId: 1 },
           { unitId: 5 },
         ]
       });
       expect(actual).toBeInstanceOf(Game);
-      expect(actual.units).toBeInstanceOf(Immutable.List);
-      expect(actual.units.get(0).unitId).toBe(1);
-      expect(actual.units.get(1).unitId).toBe(5);
+      expect(actual.field.id).toBe(2);
+      expect(actual.state.turnCount).toBe(3);
+      expect(actual.state.turn).toBeFalsy();
+      expect(actual.state.isEnd).toBeTruthy();
+      expect(actual.state.winner).toBeFalsy();
+      expect(actual.units.length).toBe(2);
+      expect(actual.units[0]).toBeInstanceOf(Unit);
+      expect(actual.units[1]).toBeInstanceOf(Unit);
+      const unitsIds = actual.units.map(u => u.status.id);
+      expect(unitsIds).toContain(1);
+      expect(unitsIds).toContain(5);
     });
   });
 
-  describe('unit()', () => {
-    test('return unit', () => {
+  describe('toData()', () => {
+    test('return object', () => {
       const game = new Game({
-        units: Immutable.List([
-          new Unit({unitId: 2, cellId: 11, hp:1}),
-          new Unit({unitId: 7, cellId: 15, hp:2}),
-        ])
+        fieldId: 2,
+        state: {
+          turnCount: 3,
+          turn: false,
+          isEnd: true,
+          winner: false,
+        },
       });
 
-      const actual = game.unit(15);
+      expect(game.toData()).toEqual({
+        fieldId: 2,
+        units: [],
+        state: {
+          turnCount: 3,
+          turn: false,
+          isEnd: true,
+          winner: false,
+        },
+      });
+    });
+  });
+
+  describe('getUnit()', () => {
+    test('return unit', () => {
+      const game = new Game({
+        units: [
+          {unitId: 2, state: {cellId:11, hp:1}},
+          {unitId: 7, state: {cellId:15, hp:2}},
+        ]
+      });
+
+      const actual = game.getUnit(15);
       expect(actual).toBeInstanceOf(Unit);
-      expect(actual.unitId).toBe(7);
+      expect(actual.status.id).toBe(7);
     });
 
     test('return falsy', () => {
       const game = new Game({
-        units: Immutable.List([
-          new Unit({unitId: 2, cellId: 11, hp:1}),
-        ])
+        units: [
+          {unitId: 2, state: {cellId: 11, hp:1}},
+        ]
       });
-      expect(game.unit(15)).toBeFalsy();
+      expect(game.getUnit(15)).toBeFalsy();
     });
   });
 
   describe('ownedUnits()', () => {
     test('return unit', () => {
       const game = new Game({
-        units: Immutable.List([
-          new Unit({unitId: 2, isOffense: true}),
-          new Unit({unitId: 4, isOffense: true}),
-          new Unit({unitId: 7, isOffense: false}),
-        ])
+        units: [
+          {unitId: 2, isOffense: true},
+          {unitId: 4, isOffense: true},
+          {unitId: 7, isOffense: false},
+        ]
       });
 
       const actual = game.ownedUnits(true);
-      expect(actual.count()).toBe(2);
-      expect(actual.get(0).unitId).toBe(2);
-      expect(actual.get(1).unitId).toBe(4);
+      expect(actual[0]).toBeInstanceOf(Unit);
+      expect(actual[1]).toBeInstanceOf(Unit);
+      expect(actual.length).toBe(2);
+      const actualIds = actual.map(u => u.status.id);
+      expect(actualIds).toContain(2);
+      expect(actualIds).toContain(4);
     });
   });
 
