@@ -1,55 +1,80 @@
-import React from 'react';
-
+// @flow
+import React, { Fragment } from 'react';
 import './style.css';
 
-function Unit({ unit }) {
+import UnitImage from '../UnitImage/index.jsx';
+
+import Unit from '../../../game/models/Unit.js';
+import Terrain from '../../../game/models/Terrain.js';
+import type { Forecast } from '../../../game/models/Game.js';
+
+function TurnEndComponent({ isMyTurn, onClickEndTurn }) {
+  return (
+    <button
+      id="cp-turnend-btn"
+      disabled={!isMyTurn}
+      onClick={() => onClickEndTurn()}
+    >
+      {isMyTurn ? (
+        <span>ターン<br/>終了</span>
+      ) : (
+        <span>相手の<br/>ターン</span>
+      )}
+    </button>
+  );
+}
+
+function UnitComponent({ unit }: { unit: ?Unit }) {
   if (!unit) {
     return null;
   }
   const status = unit.status;
   return (
     <div id="cp-unit" className={unit.isOffense ? 'offense' : 'defense'}>
-      <div id="cp-unit-avatar">
+      <div id="cp-unit-header">
         <div id="cp-unit-name">{status.name}</div>
-        <div>
+        <div id="cp-unit-hp">
           <span>HP</span>
-          <span id="hp">{unit.state.hp}</span>
-          <span id="hp-max">/ {status.hp}</span>
+          <div id="cp-unit-hp-val">
+            <span >{unit.state.hp}</span>
+            <span className="max">/{status.hp}</span>
+          </div>
         </div>
       </div>
-      <table id="cp-status-table">
-        <tbody>
-          <tr>
-            <th>力</th>
-            <td>{status.pow}</td>
-            <th>幸運</th>
-            <td>{status.luc}</td>
-          </tr>
-          <tr>
-            <th>守備</th>
-            <td>{status.dff}</td>
-            <th>命中</th>
-            <td>{status.hit}</td>
-          </tr>
-          <tr>
-            <th>信仰</th>
-            <td>{status.fth}</td>
-            <th>移動</th>
-            <td>{status.move}</td>
-          </tr>
-          <tr>
-            <th>技</th>
-            <td>{status.skl}</td>
-            <th>射程</th>
-            <td>{
-              status.min_range == status.max_range
-                ? status.min_range
-                : `${status.min_range}-${status.max_range}`
-            }</td>
-        </tr>
-      </tbody>
-    </table>
-  </div>
+      <div id="cp-unit-main">
+        <div id="cp-unit-avatar">
+          <UnitImage klassId={unit.klass.id} isOffense={unit.isOffense} />
+        </div>
+        <table id="cp-status-table">
+          <tbody>
+            <tr>
+              <th>力</th>
+              <td>{status.str}</td>
+              <th>技</th>
+              <td>{status.skl}</td>
+            </tr>
+            <tr>
+              <th>守備</th>
+              <td>{status.dff}</td>
+              <th>移動</th>
+              <td>{status.move}</td>
+            </tr>
+            <tr>
+              <th>信仰</th>
+              <td>{status.fth}</td>
+              <th>射程</th>
+              <td>
+                {
+                  status.min_range == status.max_range
+                    ? status.min_range
+                    : `${status.min_range}-${status.max_range}`
+                }
+              </td>
+            </tr>
+          </tbody>
+        </table>
+      </div>
+    </div>
   );
 }
 
@@ -57,67 +82,105 @@ function Info({ isOffense, terrain, turnRemained }) {
   return (
     <div id="cp-info">
       <div id="cp-banner" className={isOffense ? 'offense' : 'defense'}>
-      {isOffense ? (
-        <div id="cp-army">攻撃軍</div>
-      ) : (
-        <div id="cp-army">防御軍</div>
-      )}
-      <div id="cp-turn">残り{turnRemained}ターン</div>
-    </div>
-    <div id="cp-terrain">
-      {terrain &&
-          <React.Fragment>
-            <span className="name">{terrain.name}</span>
-            <span className="avoid">回避</span>
-            <span className="value">
-              {terrain.avoidance > 0 ? `+${terrain.avoidance}` : terrain.avoidance}
-            </span>
-          </React.Fragment>
-      }
-    </div>
+        {isOffense ? (
+          <div id="cp-army">攻撃軍</div>
+        ) : (
+          <div id="cp-army">防衛軍</div>
+        )}
+        { turnRemained > 1 ? (
+          <div id="cp-turn">残り<b>{turnRemained}</b>ターン</div>
+        ) : (
+          <div id="cp-turn"><b>最終ターン</b></div>
+        )}
+      </div>
+      <div id="cp-terrain">
+        {terrain &&
+            <Fragment>
+              <span className="name">{terrain.name}</span>
+              <span className="avoid">回避</span>
+              <span className="value">
+                {terrain.avoidance > 0 ? `+${terrain.avoidance}` : terrain.avoidance}
+              </span>
+            </Fragment>
+        }
+      </div>
     </div>
   );
 }
 
-function Forecast({ forecast }) {
+function percentStr(value: number) {
+  return value ? `${value}%` : null;
+}
+
+function ForecastComponent({ forecast }) {
   const mySide = (forecast.me.isOffense == forecast.tg.isOffense)
     ? 'healer' : (forecast.me.isOffense ? 'offense' : 'defense');
   const opSide = forecast.tg.isOffense ? 'offense' : 'defense';
+
+  const { me, tg } = forecast;
+
   return (
     <div id="cp-forecast">
-      <ul className={mySide}>
-        <li>{forecast.me.name}</li>
-        <li>{forecast.me.hp}</li>
-        <li>{forecast.me.val || '-'}</li>
-        <li>{forecast.me.hit || '-'}</li>
-        <li>{forecast.me.crit || '-'}</li>
-      </ul>
-      <ul>
-        <li></li>
-        <li>HP</li>
-        <li>威力</li>
-        <li>命中</li>
-        <li>必殺</li>
-      </ul>
-      <ul className={opSide}>
-        <li>{forecast.tg.name}</li>
-        <li>{forecast.tg.hp}</li>
-        <li>{forecast.tg.val || '-'}</li>
-        <li>{forecast.tg.hit || '-'}</li>
-        <li>{forecast.tg.crit || '-'}</li>
-      </ul>
+      <div id="cp-forecast-wrap">
+        <div className={['unit', mySide].join(' ')}>
+          <div>
+            <UnitImage klassId={me.klass} isOffense={me.isOffense} />
+            <div className="name">{me.name}</div>
+          </div>
+        </div>
+        <table>
+          <tbody>
+            <tr className={mySide}>
+              <td className="important">{me.hp}</td>
+              <td className="important">{me.val || '-'}</td>
+              <td>{percentStr(me.hit) ||  '-'}</td>
+              <td>{percentStr(me.crit) || '-'}</td>
+            </tr>
+            <tr>
+              <th>HP</th>
+              <th>威力</th>
+              <th>命中</th>
+              <th>会心</th>
+            </tr>
+            <tr className={opSide}>
+              <td className="important">{tg.hp}</td>
+              <td className="important">{tg.val || '-'}</td>
+              <td>{percentStr(tg.hit) || '-'}</td>
+              <td>{percentStr(tg.crit) || '-'}</td>
+            </tr>
+          </tbody>
+        </table>
+        <div className={['unit', opSide].join(' ')}>
+          <div>
+            <UnitImage klassId={tg.klass} isOffense={tg.isOffense} />
+            <div className="name">{tg.name}</div>
+          </div>
+        </div>
+      </div>
     </div>
   );
 }
 
-export default function ControlPannel({ isOffense, terrain, unit, forecast, turnRemained }) {
+export default function ControlPannel({ isOffense, terrain, unit, forecast, turnRemained, isMyTurn, onClickEndTurn }: {
+  isOffense: boolean,
+  terrain: ?Terrain,
+  unit: ?Unit,
+  forecast: ?Forecast,
+  turnRemained: number,
+  isMyTurn: boolean,
+  onClickEndTurn: void => void,
+}) {
   return (
     <div id="cp-container">
       <div id="cp-contents">
+        <TurnEndComponent
+          isMyTurn={isMyTurn}
+          onClickEndTurn={onClickEndTurn}
+        />
         {forecast ? (
-          <Forecast forecast={forecast} />
+          <ForecastComponent forecast={forecast} />
         ) : (
-          <Unit unit={unit} />
+          <UnitComponent unit={unit} />
         )}
         <Info
           isOffense={isOffense}
