@@ -14,12 +14,14 @@ export default class Match extends React.Component {
   }
 
   componentDidMount() {
+    /*
     this.popstateListener = () => {
       if (this.state.isLocked) {
         history.pushState(null, null, null);
       }
     };
     window.addEventListener('popstate', this.popstateListener);
+    */
     this.unloadListener = e => {
       if (this.state.isLocked) {
         e.returnValue = 'test';
@@ -29,20 +31,22 @@ export default class Match extends React.Component {
   }
 
   componentWillUnmount() {
+    /*
     if (this.state.isLocked) {
       history.back();
     }
     window.removeEventListener('popstate', this.popstateListener);
+    */
     window.removeEventListener('beforeunload', this.unloadListener);
   }
 
   componentWillReceiveProps(nextProps) {
     const { room } = nextProps;
     if (room && !this.state.isLocked) {
-      history.pushState(null, null, null);
+      // history.pushState(null, null, null);
       this.setState({ isLocked: true });
     } else if (!room && this.state.isLocked) {
-      history.back();
+      // history.back();
       this.setState({ isLocked: false });
     }
   }
@@ -51,7 +55,7 @@ export default class Match extends React.Component {
     const { room } = this.props;
     if (room) {
       if (room.stateIs('SELECT')) {
-        return <Selector costLimit={16} />;
+        return <Selector costLimit={20} />;
       } else if (room.stateIs('BATTLE') && room.game) {
         return <Game />;
       }
@@ -59,7 +63,4 @@ export default class Match extends React.Component {
     return <Lobby />;
   }
 
-
 }
-
-
