@@ -16,9 +16,18 @@ export default class Top extends React.Component {
     return (
       <div id="top-base">
         <div id="top-container">
-          <div id="top-navigator">
-            <button onClick={() => onClickCreateDeck() }>デッキ生成</button>
-            <Link to="match">対戦ロビー</Link>
+          <ul id="top-navigator">
+            <li>
+              <button className="top-btn" onClick={() => onClickCreateDeck() }>ランダムデッキ</button>
+            </li>
+            {deck && 
+                <li>
+                  <Link className="top-btn strong" to="match" target="sl_game">対戦ロビー</Link>
+                </li>
+            }
+          </ul>
+          <div id="top-deck">
+            <h2>YOUR DECK</h2>
             <Deck deck={deck} />
           </div>
         </div>
