@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
-import user from './user';
-import match from './match';
+import user from './user.js';
+import match from './match.js';
 
 export default combineReducers({
   user,
